Aloitettiin otteluiden luomisen muokkaus siten siten että paritskabatcontroller palauttaa pelaajalistan ja siirtyy pääikkunan kautta otteluttutloksetcontrolleriin

+++ src/htSpike/TiistaiSpike.java
@@ -31,7 +31,7 @@ public class TiistaiSpike {
     /**
      * @return palauttaa otteluiden ranking listan
      */
-    public TreeMap<Double, Integer> getRanking(){
+    public TreeMap<Integer, Double> getRanking(){
         return this.ottelut.getRanking();
     }
